add unsigned zkLogin extrinsic

+++ primitives/zklogin/src/test_helper.rs
@@ -183,7 +183,7 @@ fn gen_address_seed_with_salt_hash(
     .to_string())
 }
 
-pub fn get_raw_data() -> (AccountId32, String, u64, [u8; 32]) {
+pub fn get_raw_data() -> (AccountId32, String, u32, [u8; 32]) {
     let user_salt = "6903439401297002981078976741241818963710729444388942281949823152082404716376301797176193848";
 
     let address_seed = gen_address_seed(
@@ -232,7 +232,7 @@ pub fn get_raw_data() -> (AccountId32, String, u64, [u8; 32]) {
         "header": "913143068733459984664279033783989157259274322902058410967852973431920544493"
     }"#;
 
-    let max_epoch: u64 = 834;
+    let max_epoch: u32 = 834;
     let eph_pubkey_bytes: [u8; 32] = [
         250, 253, 29, 158, 37, 168, 126, 150, 82, 151, 106, 123, 176, 108, 46, 71, 119, 194, 229,
         57, 217, 15, 62, 231, 182, 177, 43, 154, 69, 17, 138, 136,
